requirements.txt
duckdb==0.10.0
fastavro==1.9.7
pandas==2.0.3
python-dotenv==1.0.1
Requests==2.32.3
